Move converters to bot/ This also updates all the references to `converters.py` to moved version. It makes more sense to have `converters.py` inside bot/ than bot/utils
